agent-sdk-ts: extract tool truncation helpers (oh-tab-g9ys)
